se creo las vistas, y se movieron los archivos del nav
