+++ app/src/main/java/com/example/location_feature/vistas/comentarios_cuenta.kt
@@ -1,31 +1,32 @@
-package com.example.location_feature
+package com.example.location_feature.vistas
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.location_feature.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
+private const val ARG_PARAM1_COMENTARIOS = "param11"
+private const val ARG_PARAM2_COMENTARIOS = "param22"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [home_mapa.newInstance] factory method to
+ * Use the [comentarios_cuenta.newInstance] factory method to
  * create an instance of this fragment.
  */
-class home_mapa : Fragment() {
+class comentarios_cuenta : Fragment() {
     // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
+    private var param11: String? = null
+    private var param22: String? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
+            param11 = it.getString(ARG_PARAM1_COMENTARIOS)
+            param22 = it.getString(ARG_PARAM2_COMENTARIOS)
         }
     }
 
@@ -34,7 +35,7 @@ class home_mapa : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home_mapa, container, false)
+        return inflater.inflate(R.layout.fragment_comentarios_cuenta, container, false)
     }
 
     companion object {
@@ -44,15 +45,15 @@ class home_mapa : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment home_mapa.
+         * @return A new instance of fragment comentarios_cuenta.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            home_mapa().apply {
+            comentarios_cuenta().apply {
                 arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
+                    putString(ARG_PARAM1_COMENTARIOS, param1)
+                    putString(ARG_PARAM2_COMENTARIOS, param2)
                 }
             }
     }
