+++ app/src/main/java/com/example/location_feature/vistas/detail_dialog_calendario_rutas_zonas.kt
@@ -1,10 +1,11 @@
-package com.example.location_feature
+package com.example.location_feature.vistas
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.location_feature.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -13,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [crear_cuenta.newInstance] factory method to
+ * Use the [detail_dialog_calendario_rutas_zonas.newInstance] factory method to
  * create an instance of this fragment.
  */
-class crear_cuenta : Fragment() {
+class detail_dialog_calendario_rutas_zonas : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -34,7 +35,11 @@ class crear_cuenta : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_crear_cuenta, container, false)
+        return inflater.inflate(
+            R.layout.fragment_detail_dialog_calendario_rutas_zonas,
+            container,
+            false
+        )
     }
 
     companion object {
@@ -44,12 +49,12 @@ class crear_cuenta : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment crear_cuenta.
+         * @return A new instance of fragment detail_dialog_calendario_rutas_zonas.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            crear_cuenta().apply {
+            detail_dialog_calendario_rutas_zonas().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
