+++ app/src/main/java/com/example/location_feature/vistas/home_mapa.kt
@@ -1,10 +1,11 @@
-package com.example.location_feature
+package com.example.location_feature.vistas
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.location_feature.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -13,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [editar_perfil.newInstance] factory method to
+ * Use the [home_mapa.newInstance] factory method to
  * create an instance of this fragment.
  */
-class editar_perfil : Fragment() {
+class home_mapa : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -34,7 +35,7 @@ class editar_perfil : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_editar_perfil, container, false)
+        return inflater.inflate(R.layout.fragment_home_mapa, container, false)
     }
 
     companion object {
@@ -44,12 +45,12 @@ class editar_perfil : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment editar_perfil.
+         * @return A new instance of fragment home_mapa.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            editar_perfil().apply {
+            home_mapa().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
